moved screencopy protocol under protocols
