+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod input;
 pub mod layout;
 pub mod shell;
 pub mod state;
+pub mod protocols;
 
 pub use errors::{CompositorError, Result};
 pub use state::ProjectWC;
